Fixed MaterialSideNav classes inconsistent #725
